requirements.txt
google-generativeai==0.8.6
matplotlib==3.10.8
